# AVX2 ggml_vec_dot_q4_0 performance improvement ~5%