Prefetches some data before it's usage
(basing on original work: https://github.com/ggerganov/llama.cpp/pull/295)

Before (usually 245 - 260 ms):
![prefetch_wo](https://user-images.githubusercontent.com/76458234/229905870-38a375df-e36a-4aed-a4b2-fc5db06371fe.png)

After (usually 226 - 234 ms): 
![prefetch_w](https://user-images.githubusercontent.com/76458234/229905879-00f81e7d-2d1d-4c2b-8564-82e99ee3e980.png)


Tested on Windows 10 + i7-10700k, needs further testing on different OSes and CPUs to make sure there's no unintentional issues hence draft. Also naming constants appropriately!