@rabidcopy 2023-04-04
Well, running master and this PR about 5 times in a row each, here's what I got.
`./main -m ../alpaca-7b-native.bin  -p "Building a website can be done in 10 simple steps:" -n 100 -t 6 --seed 1`
Latest master:
```
llama_print_timings:        load time =  1526.48 ms
llama_print_timings:      sample time =    66.08 ms /   100 runs   (    0.66 ms per run)
llama_print_timings: prompt eval time =  1784.33 ms /    14 tokens (  127.45 ms per token)
llama_print_timings:        eval time = 20384.87 ms /    99 runs   (  205.91 ms per run)
llama_print_timings:       total time = 22740.42 ms
```
This PR:
```
llama_print_timings:        load time =  2450.54 ms
llama_print_timings:      sample time =    78.16 ms /   100 runs   (    0.78 ms per run)
llama_print_timings: prompt eval time =  1924.37 ms /    14 tokens (  137.46 ms per token)
llama_print_timings:        eval time = 19275.24 ms /    99 runs   (  194.70 ms per run)
llama_print_timings:       total time = 21805.14 ms
```