@ivanstepanovftw 2023-04-04
Fedora Linux, AMD Ryzen 7 5800U
Orig:
```

llama_print_timings:        load time =  1045.07 ms
llama_print_timings:      sample time =    21.53 ms /    40 runs   (    0.54 ms per run)
llama_print_timings: prompt eval time =   902.05 ms /    16 tokens (   56.38 ms per token)
llama_print_timings:        eval time =  7273.96 ms /    39 runs   (  186.51 ms per run)
llama_print_timings:       total time =  8789.19 ms
```
This pr:
```
llama_print_timings:        load time =  1051.30 ms
llama_print_timings:      sample time =    19.13 ms /    40 runs   (    0.48 ms per run)
llama_print_timings: prompt eval time =   912.66 ms /    16 tokens (   57.04 ms per token)
llama_print_timings:        eval time =  7778.33 ms /    39 runs   (  199.44 ms per run)
llama_print_timings:       total time =  9298.49 ms
```