@ivanstepanovftw 2023-04-04
13B q4_0

```
llama_print_timings:        load time =  1535.26 ms
llama_print_timings:      sample time =    21.43 ms /    40 runs   (    0.54 ms per run)
llama_print_timings: prompt eval time =  1780.11 ms /    16 tokens (  111.26 ms per token)
llama_print_timings:        eval time = 12984.40 ms /    39 runs   (  332.93 ms per run)
llama_print_timings:       total time = 15470.78 ms
```

---

```
llama_print_timings:        load time =  1563.62 ms
llama_print_timings:      sample time =    19.79 ms /    40 runs   (    0.49 ms per run)
llama_print_timings: prompt eval time =  1844.18 ms /    16 tokens (  115.26 ms per token)
llama_print_timings:        eval time = 14039.39 ms /    39 runs   (  359.98 ms per run)
llama_print_timings:       total time = 16580.04 ms
```