@diimdeep 2023-04-05
7b q4_0

```
1 core, --mlock -s 1 -n 32
llama_print_timings:      sample time =    44.60 ms /    32 runs   (    1.39 ms per run)
llama_print_timings: prompt eval time =  2184.79 ms /     4 tokens (  546.20 ms per token)
llama_print_timings:        eval time = 22591.07 ms /    31 runs   (  728.74 ms per run)

after
llama_print_timings:      sample time =    42.91 ms /    32 runs   (    1.34 ms per run)
llama_print_timings: prompt eval time =  2153.29 ms /     4 tokens (  538.32 ms per token)
llama_print_timings:        eval time = 20680.65 ms /    31 runs   (  667.12 ms per run)



2 cores, --mlock -s 1 -n 32
llama_print_timings:      sample time =    42.92 ms /    32 runs   (    1.34 ms per run)
llama_print_timings: prompt eval time =  1106.27 ms /     4 tokens (  276.57 ms per token)
llama_print_timings:        eval time = 11378.82 ms /    31 runs   (  367.06 ms per run)

after
llama_print_timings:      sample time =    42.78 ms /    32 runs   (    1.34 ms per run)
llama_print_timings: prompt eval time =  1084.36 ms /     4 tokens (  271.09 ms per token)
llama_print_timings:        eval time = 10726.84 ms /    31 runs   (  346.03 ms per run)
```